dom/indexedDB/test/test_file_put_get_values.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    const READ_WRITE = "readwrite";

    const name = window.location.pathname;

    const objectStoreName = "Blobs";

    const blobData = { key: 1, blob: getRandomBlob(10000) };
    const fileData = { key: 2, file: getRandomFile("random.bin", 100000) };

    SpecialPowers.pushPrefEnv({ set: [["dom.indexedDB.dataThreshold", -1]] },
                              continueToNextStep);
    yield undefined;

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onupgradeneeded = grabEventAndContinueHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    is(event.type, "upgradeneeded", "Got correct event type");

    let db = event.target.result;
    db.onerror = errorHandler;

    db.createObjectStore(objectStoreName, { });

    event = yield undefined;

    is(event.type, "success", "Got correct event type");

    let objectStore = db.transaction([objectStoreName], READ_WRITE)
                        .objectStore(objectStoreName);
    request = objectStore.add(blobData.blob, blobData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    is(event.target.result, blobData.key, "Got correct key");

    request = objectStore.get(blobData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    verifyBlob(event.target.result, blobData.blob, 1);
    yield undefined;

    request = db.transaction([objectStoreName])
                .objectStore(objectStoreName).get(blobData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    verifyBlob(event.target.result, blobData.blob, 1);
    yield undefined;

    objectStore = db.transaction([objectStoreName], READ_WRITE)
                    .objectStore(objectStoreName);
    request = objectStore.add(fileData.file, fileData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    is(event.target.result, fileData.key, "Got correct key");

    request = objectStore.get(fileData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    verifyBlob(event.target.result, fileData.file, 2);
    yield undefined;

    request = db.transaction([objectStoreName])
                .objectStore(objectStoreName).get(fileData.key);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    verifyBlob(event.target.result, fileData.file, 2);
    yield undefined;

    is(bufferCache.length, 2, "Correct length");

    finishTest();
  }
  </script>
  <script type="text/javascript" src="file.js"></script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
